Updated the docker base container python files for the new format.
