+++ docker/README.md
@@ -199,41 +199,34 @@ The Leader Base Container performs the following steps:
 1. Wait until the requested number of workers have reported their status as READY along with their IP addresses.
 1. Create a working directory for this problem.
 1. Create and write an `input.json` with the IP addresses of the worker nodes as well as the location of the problem to the working directory
-1. Invoke the executable script located at path `/competition/solver` with a single parameter: the path to the working directory. The solver script will look for the `input.json` file in the working directory. It is also the location where solver output and error logs will be written.
+1. Invoke the executable script located at path `/competition/solver` with a single parameter: the path to the working directory. The solver script will look for the `input.json` file in the working directory. It is also the location where solver output and error logs will be written.  
+1. The return code for `/competition/solver` will determine the expected result for the solver: A return code of 10 indicates SAT, 20 indicates UNSAT, 0 indicates UNKNOWN, and all other return codes indicate an error.
 1. Upon task completion, notify all workers that the task has ended.
 
-Here is an example of the `input.json` file for parallel mallob:
+Here is an example of the `input.json` file for cloud mallob:
 
 ```text
 {
-  "formula_file": "/runder/test.cnf",
-  "worker_node_ips": ["leader"]
+  "formula_file": "/mount/efs/problem.cnf",
+  "formula_language": "DIMACS", 
+  "solver_argument_list": ["-p", "another_arg"],
+  "timeout_seconds": 10,
+  "worker_node_ips": ["192.158.1.38", "192.158.2.39", ...]
 }
 ```
 
-For the interactive example in the first section, this file is generated by the `run_parallel.sh`[runner/run_parallel.sh] script. Here is an example of `input.json` for distributed mallob:
+For the interactive example in the first section, this file is generated by the `run_parallel.sh`[runner/run_parallel.sh] script. Here is an example of `solver_out.json` for distributed mallob:
 
 ```text
 {
-  "formula_file": "/runder/test.cnf",
-  "worker_node_ips": ["leader"]
-}
-```
-
-For the interactive example, this file is generated by the `run_dist_leader.sh`[runner/run_dist_leader.sh] script
-
-The Leader Base Container waits for the `solver` script to complete, and then creates a `solver_out.json` file with the following format:
-
-```text
-{
-  "return_code": Number,   // of running the solve command
-  "result": String,        // Should be one of {"SAT", "UNSAT", "UNKNOWN", "ERROR"}
   "artifacts": {
     "stdout_path": String, // Where to find the stdout of the solve run
     "stderr_path": String  // Where to find the stderr of the solve run
   }
 }
 ```
+
+**N.B.**: The return code for `/competition/solver` will determine the satisfiability of the problem for the solver: A return code of 10 indicates SAT, 20 indicates UNSAT, 0 indicates UNKNOWN, and all other return codes indicate an error.
 
 You will adapt the Mallob `solver` script to invoke your solver and coordinates with Worker nodes. You should extend the competition base image by copying your `solver` script to a new container that is based on the Competition Base Container.
