+++ docker/satcomp-images/satcomp-leader/Dockerfile
@@ -1,11 +1,12 @@
 FROM satcomp-infrastructure:common
 RUN pip3 install flask waitress boto3 pytz polling2
-COPY --chown=ecs-user satcomp-solver-resources /opt/amazon/arg_satcomp_solver_base
+COPY --chown=ecs-user satcomp-solver-resources/ /opt/amazon/
 COPY --chown=ecs-user satcomp-leader/resources/solver /competition/solver
 COPY --chown=ecs-user satcomp-leader/resources/leader /competition/leader
 ENV PYTHONPATH=/opt/amazon:.:
 RUN chmod u+x /competition/solver
 RUN chmod u+x /competition/leader
+RUN ls /opt/amazon/arg_satcomp_solver_base
 RUN chmod u+x /opt/amazon/arg_satcomp_solver_base/leader_entrypoint.py
 RUN service ssh start
 USER ecs-user
