+++ docker/satcomp-images/satcomp-solver-resources/arg_satcomp_solver_base/node_manifest/dynamodb_manifest.py
