+++ docker/satcomp-images/satcomp-solver-resources/arg_satcomp_solver_base/s3_file_system/s3_file_system.py
