+++ docker/satcomp-images/satcomp-solver-resources/arg_satcomp_solver_base/worker/worker.py
