--- docker/satcomp-images/satcomp-solver-resources/solver/solver.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class Solver:
-    """Solver interface"""
-    def solve(self, problem_loc: str, workers: list, task_uuid: str):
-        """Interface for solve command"""
